pyproject.toml
[tool.poetry]
name = "forward-forward"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
tqdm = "^4.64.1"
torchvision = "^0.14.1"
hydra-core = "^1.3.1"
high-order-layers-torch = "^1.2.4"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
